refactor: rename apis to clouddeploy
