+++ apis/clouddeploy/v1alpha1/deliverypipeline_identity.go
