+++ config/crds/resources/apiextensions.k8s.io_v1_customresourcedefinition_deploydeliverypipelines.clouddeploy.cnrm.cloud.google.com.yaml
@@ -7,9 +7,9 @@ metadata:
   labels:
     cnrm.cloud.google.com/managed-by-kcc: "true"
     cnrm.cloud.google.com/system: "true"
-  name: deploydeliverypipelines.deploy.cnrm.cloud.google.com
+  name: deploydeliverypipelines.clouddeploy.cnrm.cloud.google.com
 spec:
-  group: deploy.cnrm.cloud.google.com
+  group: clouddeploy.cnrm.cloud.google.com
   names:
     categories:
     - gcp
